backend/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.1.0
supabase>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
openai>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
